Move client certificate field to bottom of AuthFields
