+++ app/src/main/java/com/capyreader/app/ui/accounts/LoginView.kt
@@ -1,7 +1,5 @@
 package com.capyreader.app.ui.accounts
 
-import androidx.compose.foundation.interaction.MutableInteractionSource
-import androidx.compose.foundation.interaction.PressInteraction
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
@@ -22,8 +20,6 @@ import androidx.compose.material3.Text
 import androidx.compose.material3.TextField
 import androidx.compose.material3.TopAppBar
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.LaunchedEffect
-import androidx.compose.runtime.remember
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.LocalContext
@@ -53,9 +49,9 @@ fun LoginView(
     onNavigateBack: () -> Unit = {},
     onChooseClientCert: () -> Unit = {},
     url: String,
-    clientCertAlias: String,
     username: String,
     password: String,
+    clientCertAlias: String,
     loading: Boolean = false,
     errorMessage: String? = null,
 ) {
@@ -110,23 +106,21 @@ fun LoginView(
                                     }
                                 }
                             )
-                            CertificateField(
-                                onChooseClientCert = onChooseClientCert,
-                                certAlias = clientCertAlias,
-                            )
                         }
                         AuthFields(
-                            username = username,
-                            password = password,
                             onUsernameChange = onUsernameChange,
                             onPasswordChange = onPasswordChange,
                             onSubmit = onSubmit,
+                            username = username,
+                            password = password,
                             loading = loading,
                             errorMessage = errorMessage,
-                            source = source,
                             prompt = {
                                 ServiceSignup(source)
-                            }
+                            },
+                            source = source,
+                            onChooseClientCert = onChooseClientCert,
+                            clientCertAlias = clientCertAlias,
                         )
                     }
                 }
@@ -155,34 +149,6 @@ fun UrlField(
         ),
         modifier = Modifier
             .fillMaxWidth()
-    )
-}
-
-@Composable
-fun CertificateField(
-    onChooseClientCert: () -> Unit,
-    certAlias: String,
-) {
-    TextField(
-        value = certAlias,
-        onValueChange = {},
-        singleLine = true,
-        label = {
-            Text(stringResource(R.string.auth_fields_client_certificate))
-        },
-        modifier = Modifier
-            .fillMaxWidth(),
-        readOnly = true,
-        interactionSource = remember { MutableInteractionSource() }
-            .also { interactionSource ->
-                LaunchedEffect(interactionSource) {
-                    interactionSource.interactions.collect {
-                        if (it is PressInteraction.Release) {
-                            onChooseClientCert()
-                        }
-                    }
-                }
-            }
     )
 }
