+++ app/src/main/java/com/capyreader/app/ui/articles/UpdateAuthDialog.kt
@@ -1,5 +1,6 @@
 package com.capyreader.app.ui.articles
 
+import androidx.activity.compose.LocalActivity
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.window.Dialog
@@ -15,6 +16,7 @@ fun UpdateAuthDialog(
     onSuccess: (message: String) -> Unit,
     viewModel: UpdateLoginViewModel = koinViewModel()
 ) {
+    val activity = LocalActivity.current
     val successMessage = stringResource(R.string.update_auth_success_message)
 
     Dialog(
@@ -26,6 +28,9 @@ fun UpdateAuthDialog(
                 source = viewModel.source,
                 onPasswordChange = viewModel::setPassword,
                 onNavigateBack = onDismissRequest,
+                onChooseClientCert = {
+                    activity?.let(viewModel::chooseClientCert)
+                },
                 onSubmit = {
                     viewModel.submit {
                         onSuccess(successMessage)
@@ -33,6 +38,7 @@ fun UpdateAuthDialog(
                 },
                 username = viewModel.username,
                 password = viewModel.password,
+                clientCertAlias = viewModel.clientCertAlias,
                 loading = viewModel.loading,
                 errorMessage = viewModel.errorMessage
             )
